feat: Slider Affecting RNBO Patch
baby stepping to the elevator. This version has a slider that can control the volume of the guitar volume. Everything is in place to start making this thing a bit more interesting.

For the next stage, let us do a version with a single instrument (we'll stick with guitar) but at least three ways to manipulate that sound. One of these will be using a normal Unity plugin, and one with the RNBO plugin, to see if they can work in conjunction. My theory is... they can! We shall see!