feat: Initial synth version
this version uses the simple synthesizer from the Max RNBO tutorials. It also moves the files into a dedicated RNBO folder within the project (rather than outside in the rnbo.unity.audioplugin-main project which is used to build the plugin) In the future, we should check to see if it is possible to build directly into the RNBO folder to streamling the process.

Next step - adding more signal effects to the simple synth (subtractive filter, fm synthesis? etc) using Cycling's Synth Building Blocks package as a reference.