feat: dials and starting to map them to parameters
Here is a version with 6 dials mapped to 12 sets of KeyCodes (Q/W,A/S,Z/X,E/R,D/F,C/V) that control 12 imaginary parameters. This is a helpful exercise primarily because it makes clear the difficulties there will be in mapping these parameters. Namely, if each dial has a continuous discrete value then it will be hard to map it to multiple parameters (especially if they are wildly different ranges, i.e 0-1 for volume and 400-10000 for frequency sweeps)

Need to dwell on this one for a bit...