feat: new imagery and implemented Audio
OK! Here we go! A lot of changes here. Added the AudioParam.cs scripts so that each of the dials can manipulate some sort of audio parameter. Right now we have volume and the RNBO delay (and one filter sweep), but there is the possibility for a lot more. I still want to add a distortion through RNBO because I think that can be pretty powerful/useful here.

STEP back, overall vision thoughts. This wil be one of the scenes you are tuning into with the larger synth knobs. I'd like to get that transition worked out next. What we have here is good enough for the time being, and that transition will help me to understand the overall flow of how this is going to work.