feat: fully fleshed-out main explorationSynth
Note on a weird behavior for anyone following this (or who care about making RNBO patches for Unity) it turns out that these long param names made it so that they couldn't be controlled independently in the inspector. Thus "stereo_delay_1/delay_left" and "stereo_delay_1/delay_right" were both being cut off as "stereo_delay_1/" and were intertwined. I didn't test if this would be a problem for scripting (and my theory is no) but it did make it a pain for tweaking the values during runtime.

So tl:dr keep those param names short!