fix: cleanup
is cleanup a feat? or a fix? or both!? (feels like both)

All right. A lot done here. Most recently, separated out the MainSynth script into MainSynth and LocationFinder scripts. The MainSynth was previously taking on far too much responsiblity. Rewrote a bunch of the UI_Manager so that it triggers transitions to the locations rather than y distance. Overall this will add flexibility to the whole thing.

Also did a bunch of work on the hardware side. Knob presses are working correctly. Added the indicator light, but it's not working quite right. Fix.