feat: Updated UI
Begin implementation of DR's new UI design. Overall this is nice, but the lines are probably too thin. Doesn't read from a distance. Trying to work everything into a single ParameterControl script which a) animates, and b) Updates the param, but there needs to be an animationStyle index, etc. I think this will work fine for what we're doing. Next up: finish the ParamControl script and make sure it works as expected