feat: Dials refined
the dials are refined here so that movement feels right. The graphics update in a nice way. The button click has been smoothed so it doesn't race through values (the steering wheel doesn't whiff out the car window)

The location rings are randomly mapped to dial/param combos, which is nice because then some double up, but also some are missing, and it's kind of confusing? Need to add the effects to see how it feels with audio feedback.

Note: the LockingDial is working, but backwards. Maybe this is okay, but also this makes me feel that the whole interaction is a bit boring. Maybe there is some other way to deal with this..l