feat: TextDisplay
updated the TextDisplay to be more interesting. It plays sounds, glitches, etc. It only plays the lines that are hand-coded in, so it might be more fin to have those dynamically added (and more robust), but for now, this will do.

Also rearranged the files so that all those loose scripts are back in their corral. Yee-haw.