feat: analysis pass through Meeting Notes
One that pops up here (and hasn't been used specifically anywhere else is "idea" mainly because there is a perfect example of a concrete "idea" being suggested from GV. Curious if there are other examples throughout the material of an "idea" being presented. Yes? Right? But how is that different from 'deciding' or 'options'...?